load all items from the feed
Since the feed normally returns only 100 (or 97 according to my experience?)
entries per query, the startIndex needs to be looped until everything is
downloaded once. It is accomplished by checking the changestamp of the last
item of the stream and comparing it to the largest changestamp given
by Google (included in the stream). If they are not equal, query again.